services/dialogflow-webhook/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.14.1
google-auth==2.25.2
google-api-core==2.15.0
